Sandbox Scene and Tank Fixes
This commit contains the Sandbox scene which replaces the TestingArea
and TestingAreaV2 scene.

It also contains some minor fixes to the Tank model, mainly adding
kinematic option to the RigidBody componenet of the gameobject and some
minor tweaks to the length of the capsule collider of the said
gameobject.

This helped fixing an issue with the tank death animation where the
model would float after the dying animation completes.

TankController scripts has been minorly organised in the variable
decleration section, also the health and damage variables now are public
for ease of tweaking the tank settings and for testing purposes.